# atoma-daemon/docs/openapi.yml
openapi: 3.1.0
info:
  title: atoma-daemon
  description: ''
  license:
    name: Apache-2.0
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: Local development server
paths:
...
  /nodes/claim-funds:
    post:
      tags:
      - nodes
      summary: Claims funds for completed stacks.
      operationId: nodes_claim_funds
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeClaimFundsRequest'
        required: true
      responses:
        '200':
          description: Node claim funds successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeClaimFundsResponse'
        '500':
          description: Failed to submit claim funds
  /nodes/model-subscribe:
    post:
      tags:
      - nodes
      summary: Subscribes a node to a specific model.
      operationId: nodes_model_subscribe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeModelSubscriptionRequest'
        required: true
      responses:
        '200':
          description: Node model subscription successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeModelSubscriptionResponse'
        '500':
          description: Failed to submit model subscription transaction
  /nodes/register:
    post:
      tags:
      - nodes
      summary: Registers a new node in the system.
      operationId: nodes_register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeRegistrationRequest'
        required: true
      responses:
        '200':
          description: Node registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeRegistrationResponse'
        '500':
          description: Failed to submit registration transaction
  /nodes/submit-attestations:
    post:
      tags:
      - nodes
      summary: Submits attestations for stack settlement.
      operationId: nodes_submit_attestations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeAttestationProofRequest'
        required: true
      responses:
        '200':
          description: Node attestation proof successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeAttestationProofResponse'
        '500':
          description: Failed to submit attestation proof
  /nodes/task-subscribe:
    post:
      tags:
      - nodes
      summary: Subscribes a node to a specific task.
      operationId: nodes_task_subscribe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeTaskSubscriptionRequest'
        required: true
      responses:
        '200':
          description: Node task subscription successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeTaskSubscriptionResponse'
        '500':
          description: Failed to submit task subscription transaction
  /nodes/task-unsubscribe:
    post:
      tags:
      - nodes
      summary: Unsubscribes a node from a specific task.
      operationId: nodes_task_unsubscribe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeTaskUnsubscriptionRequest'
        required: true
      responses:
        '200':
          description: Node task unsubscription successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeTaskUnsubscriptionResponse'
        '500':
          description: Failed to submit task unsubscription
  /nodes/task-update-subscription:
    post:
      tags:
      - nodes
      summary: Updates an existing task subscription for a node.
      operationId: nodes_task_update_subscription
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeTaskUpdateSubscriptionRequest'
        required: true
      responses:
        '200':
          description: Node task update subscription successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeTaskUpdateSubscriptionResponse'
        '500':
          description: Failed to submit task update subscription
  /nodes/try-settle-stacks:
    post:
      tags:
      - nodes
      summary: Attempts to settle stacks for a node.
      operationId: nodes_try_settle_stacks
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeTrySettleStacksRequest'
        required: true
      responses:
        '200':
          description: Node try settle stacks successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeTrySettleStacksResponse'
        '500':
          description: Failed to submit try settle stacks
...
components:
  schemas:
    NodeAttestationProofRequest:
      type: object
      description: |-
        Represents a request to submit a node attestation proof.

        This struct encapsulates the necessary parameters for submitting
        a node attestation proof.
      required:
      - stack_small_ids
      properties:
        gas:
          type: string
          description: |-
            Optional gas object ID.
            If not provided, the default is `None`.
        gas_budget:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas budget.
            If not provided, the default is `None`.
          minimum: 0
        gas_price:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas price.
            If not provided, the default is `None`.
          minimum: 0
        node_small_id:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional node badge ID.
            If not provided, the default is `None`.
        stack_small_ids:
          type: array
          items:
            type: integer
            format: int64
          description: The small IDs of the stacks to attest to.
    NodeAttestationProofResponse:
      type: object
      description: |-
        Represents a response to a node attestation proof request.

        This struct contains the transaction digests, which are unique
        identifiers for the transactions associated with the attestation proof.
      required:
      - tx_digests
      properties:
        tx_digests:
          type: array
          items:
            type: string
          description: The transaction digests.
    NodeClaimFundsRequest:
      type: object
      description: |-
        Represents a request to claim funds from a stack.

        This struct encapsulates the necessary parameters for claiming funds
        from a stack, including optional gas-related fields.
      required:
      - stack_small_ids
      properties:
        gas:
          type: string
          description: |-
            Optional gas object ID.
            If not provided, the default is `None`.
        gas_budget:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas budget.
            If not provided, the default is `None`.
          minimum: 0
        gas_price:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas price.
            If not provided, the default is `None`.
          minimum: 0
        node_badge_id:
          type: string
          description: |-
            Optional node badge IDs.
            If not provided, the default is `None`.
        stack_small_ids:
          type: array
          items:
            type: integer
            format: int64
          description: The small IDs of the stacks to claim funds from.
    NodeClaimFundsResponse:
      type: object
      description: |-
        Represents a response to a node claim funds request.

        This struct contains the transaction digest, which is a unique
        identifier for the transaction associated with the claim funds.
      required:
      - tx_digest
      properties:
        tx_digest:
          type: string
          description: The associated transaction digest.
    NodeModelSubscriptionRequest:
      type: object
      description: |-
        Represents a request to subscribe to a node model.

        This struct encapsulates the necessary parameters for subscribing
        to a node model, including optional gas-related fields.
      required:
      - model_name
      - echelon_id
      properties:
        echelon_id:
          type: integer
          format: int64
          description: The echelon ID associated with the subscription.
          minimum: 0
        gas:
          type: string
          description: |-
            Optional gas object ID.
            If not provided, the default is `None`.
        gas_budget:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas budget.
            If not provided, the default is `None`.
          minimum: 0
        gas_price:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas price.
            If not provided, the default is `None`.
          minimum: 0
        model_name:
          type: string
          description: The name of the model to subscribe to.
        node_badge_id:
          type: string
          description: |-
            Optional node badge ID.
            If not provided, the default is `None`.
    NodeModelSubscriptionResponse:
      type: object
      description: |-
        Represents a response to a node model subscription request.

        This struct contains the transaction digest, which is a unique
        identifier for the transaction associated with the node model subscription.
      required:
      - tx_digest
      properties:
        tx_digest:
          type: string
          description: |-
            The transaction digest.
            This is a unique identifier for the transaction.
    NodeRegistrationRequest:
      type: object
      description: |-
        Represents a request to register a node.

        This struct is used to encapsulate the necessary parameters
        for registering a node, including optional gas-related fields.
      properties:
        gas:
          type: string
          description: |-
            Optional gas object ID.
            If not provided, the default is `None`.
        gas_budget:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas budget.
            If not provided, the default is `None`.
          minimum: 0
        gas_price:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas price.
            If not provided, the default is `None`.
          minimum: 0
    NodeRegistrationResponse:
      type: object
      description: |-
        Represents a response to a node registration request.

        This struct contains the transaction digest, which is a unique
        identifier for the transaction associated with the node registration.
      required:
      - tx_digest
      properties:
        tx_digest:
          type: string
          description: |-
            The transaction digest.
            This is a unique identifier for the transaction.
...
    NodeTaskSubscriptionRequest:
      type: object
      description: |-
        Represents a request to subscribe to a node task.

        This struct encapsulates the necessary parameters for subscribing
        to a node task, including optional gas-related fields.
      required:
      - task_small_id
      - price_per_compute_unit
      - max_num_compute_units
      properties:
        gas:
          type: string
          description: |-
            Optional gas object ID.
            If not provided, the default is `None`.
        gas_budget:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas budget.
            If not provided, the default is `None`.
          minimum: 0
        gas_price:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas price.
            If not provided, the default is `None`.
          minimum: 0
        max_num_compute_units:
          type: integer
          format: int64
          description: The maximum number of compute units.
          minimum: 0
        node_badge_id:
          type: string
          description: |-
            Optional node badge ID.
            If not provided, the default is `None`.
        price_per_compute_unit:
          type: integer
          format: int64
          description: The price per compute unit.
          minimum: 0
        task_small_id:
          type: integer
          format: int64
          description: The small ID of the task to subscribe to.
    NodeTaskSubscriptionResponse:
      type: object
      description: |-
        Represents a response to a node task subscription request.

        This struct contains the transaction digest, which is a unique
        identifier for the transaction associated with the node task subscription.
      required:
      - tx_digest
      properties:
        tx_digest:
          type: string
          description: |-
            The transaction digest.
            This is a unique identifier for the transaction.
    NodeTaskUnsubscriptionRequest:
      type: object
      description: |-
        Represents a request to unsubscribe from a node task.

        This struct encapsulates the necessary parameters for unsubscribing
        from a node task, including optional gas-related fields.
      required:
      - task_small_id
      properties:
        gas:
          type: string
          description: |-
            Optional gas object ID.
            If not provided, the default is `None`.
        gas_budget:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas budget.
            If not provided, the default is `None`.
          minimum: 0
        gas_price:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas price.
            If not provided, the default is `None`.
          minimum: 0
        node_badge_id:
          type: string
          description: |-
            Optional node badge ID.
            If not provided, the default is `None`.
        task_small_id:
          type: integer
          format: int64
          description: The small ID of the task to unsubscribe from.
    NodeTaskUnsubscriptionResponse:
      type: object
      description: |-
        Represents a response to a node task unsubscription request.

        This struct contains the transaction digest, which is a unique
        identifier for the transaction associated with the node task unsubscription.
      required:
      - tx_digest
      properties:
        tx_digest:
          type: string
          description: |-
            The transaction digest.
            This is a unique identifier for the transaction.
    NodeTaskUpdateSubscriptionRequest:
      type: object
      required:
      - task_small_id
      - price_per_compute_unit
      - max_num_compute_units
      properties:
        gas:
          type: string
          description: |-
            Optional gas object ID.
            If not provided, the default is `None`.
        gas_budget:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas budget.
            If not provided, the default is `None`.
          minimum: 0
        gas_price:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas price.
            If not provided, the default is `None`.
          minimum: 0
        max_num_compute_units:
          type: integer
          format: int64
          description: The maximum number of compute units.
          minimum: 0
        node_badge_id:
          type: string
          description: |-
            Optional node badge ID.
            If not provided, the default is `None`.
        price_per_compute_unit:
          type: integer
          format: int64
          description: The price per compute unit.
          minimum: 0
        task_small_id:
          type: integer
          format: int64
          description: The small ID of the task to subscribe to.
    NodeTaskUpdateSubscriptionResponse:
      type: object
      description: |-
        Represents a response to a node task update subscription request.

        This struct contains the transaction digest, which is a unique
        identifier for the transaction associated with the node task subscription.
      required:
      - tx_digest
      properties:
        tx_digest:
          type: string
          description: |-
            The transaction digest.
            This is a unique identifier for the transaction.
    NodeTrySettleStacksRequest:
      type: object
      description: |-
        Represents a request to try settling a stack.

        This struct encapsulates the necessary parameters for attempting
        to settle a stack, including optional gas-related fields.
      required:
      - stack_small_ids
      - num_claimed_compute_units
      properties:
        gas:
          type: string
          description: |-
            Optional gas object ID.
            If not provided, the default is `None`.
        gas_budget:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas budget.
            If not provided, the default is `None`.
          minimum: 0
        gas_price:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Optional gas price.
            If not provided, the default is `None`.
          minimum: 0
        node_badge_id:
          type: string
          description: |-
            Optional node badge ID.
            If not provided, the default is `None`.
        num_claimed_compute_units:
          type: integer
          format: int64
          description: The number of compute units claimed.
          minimum: 0
        stack_small_ids:
          type: array
          items:
            type: integer
            format: int64
          description: The small IDs of the stacks to settle.
    NodeTrySettleStacksResponse:
      type: object
      description: |-
        Represents a response to a node try settle stack request.

        This struct contains the transaction digest, which is a unique
        identifier for the transaction associated with the attempt to settle a stack.
      required:
      - tx_digests
      properties:
        tx_digests:
          type: array
          items:
            type: string
          description: |-
            The transaction digests.
            This is a unique identifier for the transaction.
...